// src/model/src/lib/enums/muscle.ts
export enum Muscle {
    QUADS = "quads",
    HAMSTRINGS = "hamstrings",
    GLUTES = "glutes",
    CALVES = "calves",
    HIP_FLEXORS = "hip flexors",  
    SHOULDERS = "shoulders",
    LATS = "lats",
    TRICEPS = "triceps",
    FOREARMS = "forearms",
    WRISTS = "wrists",
    ROTATOR_CUFF = "rotator cuff",
    NECK = "neck",
    UPPER_BODY = "upper body", 
    CORE = "core",
    CORE_STABILITY = "core stability",
    LOWER_BACK = "lower back",
    FAST_TWITCH_FIBERS = "fast-twitch fibers",
    SLOW_TWITCH_FIBERS = "slow-twitch fibers",
    CARDIOVASCULAR_SYSTEM = "cardiovascular system",
    RESPIRATORY_SYSTEM = "respiratory muscles",
    AEROBIC_SYSTEM = "aerobic system",
    ANAEROBIC_SYSTEM = "anaerobic system",
    LACTATE_THRESHOLD = "lactate threshold",
    LACTATE_CLEARANCE = "lactate clearance system",
    VO2_MAX = "VO2 max",
    ANAEROBIC_CAPACITY = "anaerobic capacity",
    NEUROMUSCULAR_COORDINATION = "neuromuscular coordination",
    COORDINATION = "coordination",
    PROPRIOCEPTION = "proprioception",
    EXPLOSIVE_POWER = "explosive power",
    POWER = "power",
    SPEED_ENDURANCE = "speed endurance",
    AEROBIC_ENDURANCE = "aerobic endurance",
    MENTAL_TOUGHNESS = "mental toughness",
    MENTAL_ENDURANCE = "mental endurance",
    MENTAL_ADAPTABILITY = "mental adaptability",
    PEDALING_TECHNIQUE = "pedaling technique",
    RUNNING_ECONOMY = "running economy",
    RUNNING_FORM = "running form",
    PACING_CONTROL = "pacing control",
    MARATHON_SPECIFIC_FITNESS = "marathon-specific fitness",
    FULL_BODY = "full body",
    FULL_LOWER_BODY = "full lower body"
}